+++ seccamera/external/libexifa/pentax/exif-mnote-data-pentax.c
@@ -91,7 +91,7 @@ exif_mnote_data_pentax_save (ExifMnoteData *ne,
 	 * Allocate enough memory for header, the number of entries, entries,
 	 * and next IFD pointer
 	 */
-	*buf_size = o2 + 2 + n->count * 12 + 4;
+	*buf_size = (unsigned int)(o2 + 2 + n->count * 12 + 4);
 	switch (n->version) {
 	case casioV2:
 		base = MNOTE_PENTAX2_TAG_BASE;
@@ -187,7 +187,7 @@ exif_mnote_data_pentax_save (ExifMnoteData *ne,
 				return;
 			}
 			*buf = t;
-			*buf_size = ts;
+			*buf_size = (unsigned int)ts;
 			exif_set_long (*buf + o, n->order, datao + doff);
 		} else
 			doff = o;
@@ -299,7 +299,7 @@ exif_mnote_data_pentax_load (ExifMnoteData *en,
 		 */
 		s = exif_format_get_size (n->entries[tcount].format) *
                                       n->entries[tcount].components;
-		n->entries[tcount].size = s;
+		n->entries[tcount].size = (unsigned int)s;
 		if (s) {
 			size_t dataofs = o + 8;
 			if (s > 4)
@@ -325,7 +325,7 @@ exif_mnote_data_pentax_load (ExifMnoteData *en,
 		++tcount;
 	}
 	/* Store the count of successfully parsed tags */
-	n->count = tcount;
+	n->count = (unsigned int)tcount;
 }
 
 static unsigned int
